Rename adapter metrics module to common
