+++ application/src/main/java/com/xavelo/sqs/adapter/in/http/quote/QuoteController.java
@@ -1,7 +1,7 @@
 package com.xavelo.sqs.adapter.in.http.quote;
 
-import com.xavelo.adaptermetrics.Adapter;
-import com.xavelo.adaptermetrics.CountAdapterInvocation;
+import com.xavelo.common.metrics.Adapter;
+import com.xavelo.common.metrics.CountAdapterInvocation;
 import com.xavelo.sqs.adapter.in.http.quote.mapper.QuoteMapper;
 import com.xavelo.sqs.application.api.QuoteApi;
 import com.xavelo.sqs.application.api.model.QuoteDto;
@@ -18,8 +18,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-import static com.xavelo.adaptermetrics.AdapterMetrics.Direction.IN;
-import static com.xavelo.adaptermetrics.AdapterMetrics.Type.HTTP;
+import static com.xavelo.common.metrics.AdapterMetrics.Direction.IN;
+import static com.xavelo.common.metrics.AdapterMetrics.Type.HTTP;
 
 @Adapter
 @RestController
