+++ common/src/main/java/com/xavelo/common/metrics/autoconfigure/AdapterMetricsAutoConfiguration.java
@@ -1,6 +1,6 @@
-package com.xavelo.adaptermetrics.autoconfigure;
+package com.xavelo.common.metrics.autoconfigure;
 
-import com.xavelo.adaptermetrics.AdapterMetricsAspect;
+import com.xavelo.common.metrics.AdapterMetricsAspect;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
